Made tf-state bucket deletion optional, defaulting to false. (#29)
* Made tf-state bucket deletion optional, defaulting to false.

* Destroy bucket only when TERRAFORM_DESTROY & TF_STATE_BUCKET_DESTROY = true

* Updating `tf_state_bucket_destroy` descriptions

* Adding warning per Mick suggestion

* Typo

+++ operations/deployment/terraform/bitops.after-deploy.d/delete-tf-state-bucket.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-if [[ $TERRAFORM_DESTROY = true ]] ; then
-  echo "Destroying S3 buket"
+if [[ $TERRAFORM_DESTROY = true ]] && [[ $TF_STATE_BUCKET_DESTROY = true ]]; then
+  echo "Destroying S3 buket --> $TF_STATE_BUCKET"
   aws s3 rb s3://$TF_STATE_BUCKET --force
+else 
+  if [[ $TERRAFORM_DESTROY != true ]] && [[ $TF_STATE_BUCKET_DESTROY = true ]]; then
+    echo "TF_STATE_BUCKET_DESTROY set to true, but TERRAFORM_DESTROY wasn't. Not destroying the state bucket ($TF_STATE_BUCKET)"
+  fi
 fi
